requirements.txt
flask
pillow
python-docx
PyPDF2
reportlab